# backend-core/swagger.yaml
openapi: 3.0.2
info:
  title: Boards API
  description: Boards public API documentation
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.17
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /users:
    post:
      tags:
        - users
      summary: Create user
      description: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                is_guest:
                  type: boolean
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - name
                - is_guest
      responses:
        '201':
          description: Successfully created user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/me:
    get:
      tags:
        - users
      summary: Get user
      description: Get user details using auth bearer token
      responses:
        '200':
          description: Successfully retrieved user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - bearerAuth: []
  /users/search:
    get:
      tags:
        - users
      summary: List users
      description: Retrieve a list of top 10 users based on query parameter
      parameters:
        - name: email
          in: query
          description: Search for users using email
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
  /users/email-verifications:
    post:
      tags:
        - users
      summary: Create a email verification record
      description: Create a email verification record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                code:
                  type: string
              required:
                - user_id
                - code
      responses:
        '201':
          description: Successfully created email verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailVerification'
  /users/verify-email:
    post:
      tags:
        - users
      summary: Create a email verification record
      description: Create a email verification record
      parameters:
        - name: user_id
          in: query
          description: ID of user
          required: true
          schema:
            type: string
        - name: code
          in: query
          description: Email verification code
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Successfully verified email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: Email verified.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID of the user
        name:
          type: string
          maxLength: 255
          description: Name of the user
        email:
          type: string
          format: email
          maxLength: 255
          uniqueItems: true
          description: Email of the user
        password:
          type: string
          maxLength: 255
          description: Password of the user
        is_guest:
          type: boolean
          default: false
          description: Flag indicating if the user is a guest
        created_at:
          type: string
          format: date-time
          description: Timestamp indicating the creation date
        updated_at:
          type: string
          format: date-time
          description: Timestamp indicating the last update date
      example:
        id: a5f4c43f-9ae3-4e6d-9682-97351cbb864e
        name: John Doe
        email: johndoe@example.com
        password: password123
        is_guest: false
        created_at: '2023-06-12T10:00:00Z'
        updated_at: '2023-06-12T10:00:00Z'
    EmailVerification:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID of the email verification
        user_id:
          type: string
          format: uuid
          description: ID of the user
        is_verified:
          type: boolean
          default: false
          description: Verification status of the record
        created_at:
          type: string
          format: date-time
          description: Timestamp indicating the creation date
        updated_at:
          type: string
          format: date-time
          description: Timestamp indicating the last update date
      example:
        id: 426d0f9e-dbbd-4859-884c-276175c94ee3
        user_id: d0865843-8494-4d6a-b9be-5c8f7d0e568f
        is_verified: null
        created_at: '2023-06-12T10:00:00Z'
        updated_at: '2023-06-12T10:00:00Z'
...